Cleanup, fix dead code warnings
See https://github.com/rust-lang/rust/issues/46379.
