+++ Cargo.toml
@@ -89,4 +89,3 @@ xdg = "2.4"
 
 [dev-dependencies]
 assert_cmd = "2"
-delegate = "0.6"
